pyproject.toml
[tool.poetry]
name = "internet_speedtest"
version = "0.1.0"
description = ""
authors = ["Hamzah Iqbal <hamzahiqbal.acc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
speedtest-cli = "^2.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
